+++ src/main/java/prgms/vouchermanagementapp/io/Writer.java
@@ -1,11 +1,11 @@
 package prgms.vouchermanagementapp.io;
 
 import org.springframework.stereotype.Component;
+import prgms.vouchermanagementapp.storage.model.FileVoucherRecord;
+import prgms.vouchermanagementapp.storage.model.MemoryVoucherRecord;
+import prgms.vouchermanagementapp.storage.model.VoucherRecord;
 import prgms.vouchermanagementapp.voucher.VoucherType;
 import prgms.vouchermanagementapp.voucher.model.Voucher;
-import prgms.vouchermanagementapp.voucher.warehouse.model.FileVoucherRecord;
-import prgms.vouchermanagementapp.voucher.warehouse.model.MemoryVoucherRecord;
-import prgms.vouchermanagementapp.voucher.warehouse.model.VoucherRecord;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -110,6 +110,8 @@ public class Writer {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+
+//        Objects.requireNonNull()
     }
 
     public void printMemoryVoucherRecord(List<Voucher> list) {
