+++ src/main/java/prgms/vouchermanagementapp/io/model/Amount.java
@@ -1,4 +1,4 @@
-package prgms.vouchermanagementapp.model;
+package prgms.vouchermanagementapp.io.model;
 
 import java.util.Objects;
