+++ src/main/java/prgms/vouchermanagementapp/storage/MemoryVouchers.java
@@ -1,10 +1,10 @@
-package prgms.vouchermanagementapp.voucher.warehouse;
+package prgms.vouchermanagementapp.storage;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
+import prgms.vouchermanagementapp.storage.model.MemoryVoucherRecord;
+import prgms.vouchermanagementapp.storage.model.VoucherRecord;
 import prgms.vouchermanagementapp.voucher.model.Voucher;
-import prgms.vouchermanagementapp.voucher.warehouse.model.MemoryVoucherRecord;
-import prgms.vouchermanagementapp.voucher.warehouse.model.VoucherRecord;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -13,7 +13,7 @@ import java.util.UUID;
 
 @Component
 @Profile("dev")
-public class MemoryVouchers implements VoucherWarehouse {
+public class MemoryVouchers implements Vouchers {
 
     private static final Map<UUID, Voucher> vouchers = new HashMap<>();
