+++ src/main/java/prgms/vouchermanagementapp/storage/model/FileVoucherRecord.java
@@ -1,4 +1,4 @@
-package prgms.vouchermanagementapp.voucher.warehouse.model;
+package prgms.vouchermanagementapp.storage.model;
 
 public class FileVoucherRecord implements VoucherRecord {
