+++ src/main/java/prgms/vouchermanagementapp/storage/model/VoucherRecord.java
@@ -1,0 +1,6 @@
+package prgms.vouchermanagementapp.storage.model;
+
+public interface VoucherRecord {
+
+    boolean isFile();
+}
