--- src/main/java/prgms/vouchermanagementapp/voucher/warehouse/VoucherWarehouse.java
@@ -1,11 +1,0 @@
-package prgms.vouchermanagementapp.voucher.warehouse;
-
-import prgms.vouchermanagementapp.voucher.model.Voucher;
-import prgms.vouchermanagementapp.voucher.warehouse.model.VoucherRecord;
-
-public interface VoucherWarehouse {
-
-    void store(Voucher voucher);
-
-    VoucherRecord getVoucherRecord();
-}
