--- src/main/java/prgms/vouchermanagementapp/voucher/warehouse/model/VoucherRecord.java
@@ -1,6 +1,0 @@
-package prgms.vouchermanagementapp.voucher.warehouse.model;
-
-public interface VoucherRecord {
-
-    boolean isFile();
-}
