+++ src/test/java/prgms/vouchermanagementapp/voucher/warehouse/FileVouchersTest.java
@@ -7,7 +7,8 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Profile;
 import prgms.vouchermanagementapp.VoucherManagementApp;
-import prgms.vouchermanagementapp.model.Ratio;
+import prgms.vouchermanagementapp.io.model.Ratio;
+import prgms.vouchermanagementapp.storage.Vouchers;
 import prgms.vouchermanagementapp.voucher.VoucherCreationFactory;
 import prgms.vouchermanagementapp.voucher.model.Voucher;
 
@@ -19,7 +20,7 @@ class FileVouchersTest {
     private static final String FILE_PATH = "src/main/resources/";
 
     ApplicationContext applicationContext = new AnnotationConfigApplicationContext(VoucherManagementApp.class);
-    VoucherWarehouse voucherWarehouse = applicationContext.getBean(VoucherWarehouse.class);
+    Vouchers vouchers = applicationContext.getBean(Vouchers.class);
 
     @DisplayName("파일이 지정한 위치에 정상 생성 되는지 확인")
     @Test
@@ -29,7 +30,7 @@ class FileVouchersTest {
 
         // when
         Voucher voucher = VoucherCreationFactory.createVoucher(new Ratio(3));
-        voucherWarehouse.store(voucher);
+        vouchers.store(voucher);
 
         // then
         Assertions.assertThat(file.exists()).isTrue();
